Issue #151 Removed two longs from base library since we don’t support longs. Also added empty implementations for ATHROW, MONITORENTER, and MONITOREXIT. The MONITOR opcodes won’t be necessary if we switch to a more cooperative thread model. We may need to fix ATHROW.

+++ src/lib/base/java/java/lang/String.java
@@ -646,8 +646,8 @@ public final class String
 		int po = other.offset + ooffset;
 
 		// Note: toffset, ooffset, or len might be near -1>>>1.
-		if ((ooffset < 0) || (toffset < 0) || (toffset > (long) count - len)
-				|| (ooffset > (long) other.count - len))
+		if ((ooffset < 0) || (toffset < 0) || (toffset > count - len)
+				|| (ooffset > other.count - len))
 		{
 			return false;
 		}
